Move project to Poetry
