+++ .github/funding.yml
@@ -1,5 +1,5 @@
 # These are supported funding model platforms
 ---
 github: jtprogru
-patreon: jtprog
+patreon: jtprogru
 custom: [ "https://sobe.ru/na/jtprogru" ]
