--- dev-requirements.in
@@ -1,11 +1,0 @@
-autopep8
-coverage
-flake8
-flake8-colors
-pytest
-black
-mccabe
-mypy
-pylint
-tox
-tox-gh-actions
