--- dev-requirements.txt
@@ -1,42 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile dev-requirements.in
-#
-attrs==20.3.0
-    # via pytest
-autopep8==1.5.6
-    # via -r dev-requirements.in
-coverage==5.5
-    # via -r dev-requirements.in
-flake8-colors==0.1.9
-    # via -r dev-requirements.in
-flake8==3.9.0
-    # via
-    #   -r dev-requirements.in
-    #   flake8-colors
-iniconfig==1.1.1
-    # via pytest
-mccabe==0.6.1
-    # via flake8
-packaging==20.9
-    # via pytest
-pluggy==0.13.1
-    # via pytest
-py==1.10.0
-    # via pytest
-pycodestyle==2.7.0
-    # via
-    #   autopep8
-    #   flake8
-pyflakes==2.3.1
-    # via flake8
-pyparsing==2.4.7
-    # via packaging
-pytest==6.2.3
-    # via -r dev-requirements.in
-toml==0.10.2
-    # via
-    #   autopep8
-    #   pytest
