--- requirements.in
@@ -1,1 +1,0 @@
-twitter
