--- requirements.txt
@@ -1,8 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile requirements.in
-#
-twitter==1.18.0
-    # via -r requirements.in
