+++ twtrexcs/__init__.py
@@ -1,0 +1,3 @@
+# coding=utf-8
+
+__version__ = "2021.1.0"
