+++ twtrexcs/__main__.py
@@ -14,36 +14,8 @@ import os
 import twitter
 
 from twtrexcs.helpers import get_excuse
-
-
-class TwtrExcsException(Exception):
-    pass
-
-
-class CustomFormatter(logging.Formatter):
-    """
-    Colored Formatter for logging
-    """
-
-    grey = "\x1b[38;21m"
-    yellow = "\x1b[33;21m"
-    red = "\x1b[31;21m"
-    bold_red = "\x1b[31;1m"
-    reset = "\x1b[0m"
-    format_string = "%(asctime)s - %(name)s - %(levelname)s - (%(filename)s:%(lineno)d) - %(message)s"
-
-    FORMATS = {
-        logging.DEBUG: grey + format_string + reset,
-        logging.INFO: grey + format_string + reset,
-        logging.WARNING: yellow + format_string + reset,
-        logging.ERROR: red + format_string + reset,
-        logging.CRITICAL: bold_red + format_string + reset,
-    }
-
-    def format(self, record):
-        log_fmt = self.FORMATS.get(record.levelno)
-        formatter = logging.Formatter(log_fmt)
-        return formatter.format(record)
+from twtrexcs.exceptions import TwtrExcsException
+from twtrexcs.log_formatter import TwtrFormatter
 
 
 logger = logging.getLogger(__name__)
@@ -53,7 +25,7 @@ logger.setLevel(logging.DEBUG)
 fh = logging.FileHandler(os.getenv("LOG_FILE_PATH", "twitterbot.log"))
 fh.setLevel(logging.DEBUG)
 
-fh.setFormatter(CustomFormatter())
+fh.setFormatter(TwtrFormatter())
 
 logger.addHandler(fh)
