+++ twtrexcs/exceptions.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+class TwtrExcsException(Exception):
+    pass
